# improve performance of parseRange