--- a/lib/SendStream.js
+++ b/lib/SendStream.js
@@ -132,7 +132,7 @@ function SendStream (req, path, options) {
   this._maxage = typeof this._maxage === 'string'
     ? ms.parse(this._maxage)
     : Number(this._maxage)
-  this._maxage = !isNaN(this._maxage)
+  this._maxage = this._maxage === this._maxage // eslint-disable-line no-self-compare
     ? Math.min(Math.max(0, this._maxage), MAX_MAXAGE)
     : 0
 
@@ -328,10 +328,10 @@ SendStream.prototype.isPreconditionFailure = function isPreconditionFailure () {
   if ('if-unmodified-since' in req.headers) {
     const ifUnmodifiedSince = req.headers['if-unmodified-since']
     const unmodifiedSince = Date.parse(ifUnmodifiedSince)
-    if (!Number.isNaN(unmodifiedSince)) {
+    if (unmodifiedSince === unmodifiedSince) { // eslint-disable-line no-self-compare
       const lastModified = Date.parse(res.getHeader('Last-Modified'))
       if (
-        Number.isNaN(lastModified) ||
+        lastModified !== lastModified || // eslint-disable-line no-self-compare
         lastModified > unmodifiedSince
       ) {
         return true
@@ -441,14 +441,17 @@ SendStream.prototype.isRangeFresh = function isRangeFresh () {
   }
 
   const ifRangeTimestamp = Date.parse(ifRange)
-  if (Number.isNaN(ifRangeTimestamp)) {
+  if (ifRangeTimestamp !== ifRangeTimestamp) { // eslint-disable-line no-self-compare
     return false
   }
 
   // if-range as modified date
   const lastModified = Date.parse(this.res.getHeader('Last-Modified'))
 
-  return Number.isNaN(lastModified) || lastModified <= ifRangeTimestamp
+  return (
+    lastModified !== lastModified || // eslint-disable-line no-self-compare
+    lastModified <= ifRangeTimestamp
+  )
 }
 
 /**
@@ -586,7 +589,6 @@ SendStream.prototype.send = function send (path, stat) {
   const opts = {}
   const res = this.res
   const req = this.req
-  let ranges = req.headers.range
   let offset = options.start || 0
 
   if (res.headersSent) {
@@ -624,40 +626,44 @@ SendStream.prototype.send = function send (path, stat) {
   }
 
   // Range support
-  if (this._acceptRanges && ranges !== undefined && BYTES_RANGE_REGEXP.test(ranges)) {
-    // If-Range support
-    if (!this.isRangeFresh()) {
-      debug('range stale')
-      ranges = -2
-    } else {
-      // parse
-      ranges = parseRange(len, ranges)
-    }
+  if (this._acceptRanges) {
+    const rangeHeader = req.headers.range
 
-    // unsatisfiable
-    if (ranges === -1) {
-      debug('range unsatisfiable')
+    if (
+      rangeHeader !== undefined &&
+      BYTES_RANGE_REGEXP.test(rangeHeader)
+    ) {
+      // If-Range support
+      if (this.isRangeFresh()) {
+        // parse
+        const ranges = parseRange(len, rangeHeader)
 
-      // Content-Range
-      res.setHeader('Content-Range', contentRange('bytes', len))
+        // unsatisfiable
+        if (ranges.length === 0) {
+          debug('range unsatisfiable')
 
-      // 416 Requested Range Not Satisfiable
-      return this.error(416, {
-        headers: { 'Content-Range': res.getHeader('Content-Range') }
-      })
-    }
+          // Content-Range
+          res.setHeader('Content-Range', contentRange('bytes', len))
 
-    // valid (syntactically invalid/multiple ranges are treated as a regular response)
-    if (ranges !== -2 && ranges.length === 1) {
-      debug('range %j', ranges)
+          // 416 Requested Range Not Satisfiable
+          return this.error(416, {
+            headers: { 'Content-Range': res.getHeader('Content-Range') }
+          })
+          // valid (syntactically invalid/multiple ranges are treated as a regular response)
+        } else if (ranges.length === 1) {
+          debug('range %j', ranges)
 
-      // Content-Range
-      res.statusCode = 206
-      res.setHeader('Content-Range', contentRange('bytes', len, ranges[0]))
+          // Content-Range
+          res.statusCode = 206
+          res.setHeader('Content-Range', contentRange('bytes', len, ranges[0]))
 
-      // adjust for requested range
-      offset += ranges[0].start
-      len = ranges[0].end - ranges[0].start + 1
+          // adjust for requested range
+          offset += ranges[0].start
+          len = ranges[0].end - ranges[0].start + 1
+        }
+      } else {
+        debug('range stale')
+      }
     }
   }