--- a/lib/SendStream.js
+++ b/lib/SendStream.js
@@ -26,7 +26,7 @@ const { createHtmlDocument } = require('./createHtmlDocument')
 const { createHttpError } = require('./createHttpError')
 const { isUtf8MimeType } = require('./isUtf8MimeType')
 const { normalizeList } = require('./normalizeList')
-const { parseRange } = require('./parseRange')
+const { parseBytesRange } = require('./parseBytesRange')
 const { parseTokenList } = require('./parseTokenList')
 const { setHeaders } = require('./setHeaders')
 
@@ -636,7 +636,7 @@ SendStream.prototype.send = function send (path, stat) {
       // If-Range support
       if (this.isRangeFresh()) {
         // parse
-        const ranges = parseRange(len, rangeHeader)
+        const ranges = parseBytesRange(len, rangeHeader)
 
         // unsatisfiable
         if (ranges.length === 0) {