--- a/lib/SendStream.js
+++ b/lib/SendStream.js
@@ -132,7 +132,8 @@ function SendStream (req, path, options) {
   this._maxage = typeof this._maxage === 'string'
     ? ms.parse(this._maxage)
     : Number(this._maxage)
-  this._maxage = this._maxage === this._maxage // eslint-disable-line no-self-compare
+  // eslint-disable-next-line no-self-compare
+  this._maxage = this._maxage === this._maxage // fast path of isNaN(number)
     ? Math.min(Math.max(0, this._maxage), MAX_MAXAGE)
     : 0
 
@@ -328,10 +329,12 @@ SendStream.prototype.isPreconditionFailure = function isPreconditionFailure () {
   if ('if-unmodified-since' in req.headers) {
     const ifUnmodifiedSince = req.headers['if-unmodified-since']
     const unmodifiedSince = Date.parse(ifUnmodifiedSince)
-    if (unmodifiedSince === unmodifiedSince) { // eslint-disable-line no-self-compare
+    // eslint-disable-next-line no-self-compare
+    if (unmodifiedSince === unmodifiedSince) { // fast path of isNaN(number)
       const lastModified = Date.parse(res.getHeader('Last-Modified'))
       if (
-        lastModified !== lastModified || // eslint-disable-line no-self-compare
+        // eslint-disable-next-line no-self-compare
+        lastModified !== lastModified ||// fast path of isNaN(number)
         lastModified > unmodifiedSince
       ) {
         return true
@@ -441,7 +444,8 @@ SendStream.prototype.isRangeFresh = function isRangeFresh () {
   }
 
   const ifRangeTimestamp = Date.parse(ifRange)
-  if (ifRangeTimestamp !== ifRangeTimestamp) { // eslint-disable-line no-self-compare
+  // eslint-disable-next-line no-self-compare
+  if (ifRangeTimestamp !== ifRangeTimestamp) { // fast path of isNaN(number)
     return false
   }
 
@@ -449,7 +453,8 @@ SendStream.prototype.isRangeFresh = function isRangeFresh () {
   const lastModified = Date.parse(this.res.getHeader('Last-Modified'))
 
   return (
-    lastModified !== lastModified || // eslint-disable-line no-self-compare
+    // eslint-disable-next-line no-self-compare
+    lastModified !== lastModified || // fast path of isNaN(number)
     lastModified <= ifRangeTimestamp
   )
 }