--- a/lib/parseBytesRange.js
+++ b/lib/parseBytesRange.js
@@ -17,7 +17,7 @@
  * @public
  */
 
-function parseRange (size, str) {
+function parseBytesRange (size, str) {
   // split the range string
   const values = str.slice(str.indexOf('=') + 1)
   const ranges = []
@@ -126,4 +126,4 @@ function sortByRangeStart (a, b) {
   return a.start - b.start
 }
 
-module.exports.parseRange = parseRange
+module.exports.parseBytesRange = parseBytesRange