--- a/lib/parseBytesRange.js
+++ b/lib/parseBytesRange.js
@@ -39,11 +39,13 @@ function parseBytesRange (size, str) {
     end = parseInt(values.slice(dashIdx + 1, commaIdx), 10)
 
     // -nnn
-    if (start !== start) { // eslint-disable-line no-self-compare
+    // eslint-disable-next-line no-self-compare
+    if (start !== start) { // fast path of isNaN(number)
       start = size - end
       end = size - 1
       // nnn-
-    } else if (end !== end) { // eslint-disable-line no-self-compare
+    // eslint-disable-next-line no-self-compare
+    } else if (end !== end) { // fast path of isNaN(number)
       end = size - 1
       // limit last-byte-pos to current length
     } else if (end > size - 1) {
@@ -52,8 +54,10 @@ function parseBytesRange (size, str) {
 
     // add range only on valid ranges
     if (
-      start === start && // eslint-disable-line no-self-compare
-      end === end && // eslint-disable-line no-self-compare
+      // eslint-disable-next-line no-self-compare
+      start === start && // fast path of isNaN(number)
+      // eslint-disable-next-line no-self-compare
+      end === end && // fast path of isNaN(number)
       start > -1 &&
       start <= end
     ) {