--- a/lib/parseRange.js
+++ b/lib/parseRange.js
@@ -18,102 +18,94 @@
  */
 
 function parseRange (size, str) {
-  const index = str.indexOf('=')
-
   // split the range string
-  const arr = str.slice(index + 1).split(',')
+  const values = str.slice(str.indexOf('=') + 1)
   const ranges = []
 
+  const len = values.length
+  let i = 0
+  let il = 0
+  let j = 0
+  let start
+  let end
+  let commaIdx = values.indexOf(',')
+  let dashIdx = values.indexOf('-')
+  let prevIdx = -1
+
   // parse all ranges
-  for (let i = 0; i < arr.length; i++) {
-    const range = arr[i].split('-')
-    let start = parseInt(range[0], 10)
-    let end = parseInt(range[1], 10)
+  while (true) {
+    commaIdx === -1 && (commaIdx = len)
+    start = parseInt(values.slice(prevIdx + 1, dashIdx), 10)
+    end = parseInt(values.slice(dashIdx + 1, commaIdx), 10)
 
     // -nnn
-    if (isNaN(start)) {
+    if (start !== start) { // eslint-disable-line no-self-compare
       start = size - end
       end = size - 1
       // nnn-
-    } else if (isNaN(end)) {
+    } else if (end !== end) { // eslint-disable-line no-self-compare
+      end = size - 1
+      // limit last-byte-pos to current length
+    } else if (end > size - 1) {
       end = size - 1
     }
 
-    // limit last-byte-pos to current length
-    if (end > size - 1) {
-      end = size - 1
+    // add range only on valid ranges
+    if (
+      start === start && // eslint-disable-line no-self-compare
+      end === end && // eslint-disable-line no-self-compare
+      start > -1 &&
+      start <= end
+    ) {
+      // add range
+      ranges.push({
+        start,
+        end,
+        index: j++
+      })
     }
 
-    // invalid or unsatisfiable
-    if (isNaN(start) || isNaN(end) || start > end || start < 0) {
-      continue
+    if (commaIdx === len) {
+      break
     }
-
-    // add range
-    ranges.push({
-      start,
-      end
-    })
+    prevIdx = commaIdx++
+    dashIdx = values.indexOf('-', commaIdx)
+    commaIdx = values.indexOf(',', commaIdx)
   }
 
-  if (ranges.length === 0) {
-    // unsatisifiable
-    return -1
+  // unsatisfiable
+  if (
+    j < 2
+  ) {
+    return ranges
   }
 
-  const ordered = ranges.map(mapWithIndex).sort(sortByRangeStart)
+  ranges.sort(sortByRangeStart)
 
-  let j = 0
-  const il = ordered.length
-  for (let i = 1; i < il; ++i) {
-    const range = ordered[i]
-    const current = ordered[j]
+  il = j
+  j = 0
+  i = 1
+  while (i < il) {
+    const range = ranges[i++]
+    const current = ranges[j]
 
     if (range.start > current.end + 1) {
       // next range
-      ordered[++j] = range
+      ranges[++j] = range
     } else if (range.end > current.end) {
       // extend range
       current.end = range.end
-      current.index = Math.min(current.index, range.index)
+      current.index > range.index && (current.index = range.index)
     }
   }
 
   // trim ordered array
-  ordered.length = j + 1
+  ranges.length = j + 1
 
   // generate combined range
-  const combined = ordered.sort(sortByRangeIndex).map(mapWithoutIndex)
+  ranges.sort(sortByRangeIndex)
 
-  // copy ranges type
-  combined.type = str.slice(0, index)
-
-  return combined
-}
-
-/**
- * Map function to add index value to ranges.
- * @private
- */
-
-function mapWithIndex (range, index) {
-  return {
-    start: range.start,
-    end: range.end,
-    index
-  }
-}
-
-/**
- * Map function to remove index value from ranges.
- * @private
- */
-
-function mapWithoutIndex (range) {
-  return {
-    start: range.start,
-    end: range.end
-  }
+  return ranges
 }
 
 /**