@Uzlopak 2023-02-03
@climba03003 There is a regex check so that only ranges with "bytes=" were processed. So if "none" is set, it wont be processed by parseRange.

So we can savely drop the type as we never process the "none"-case?!