@climba03003 2023-02-03
> There is a regex check so that only ranges with "bytes=" were processed. So if "none" is set, it wont be processed by parseRange.

Yes, the `.type` is not used in anywhere. 
But it is more complete for the function itself to provide the whole view of `Range` header.