@climba03003 2023-02-03
> Can I just simply rename it to parseBytesRange?

Yes